Legg til rette for flytting ved å leggje modellen så han er tilgjengeleg, og ved å sende med eit midlertidig flagg for å leggje ved pdf-vedlegg, så vi ikkje i overgangsperioden gjer det både i brevbaker og i skribenten
