+++ brevbaker/api-model-common/gradle.properties
@@ -1,1 +1,1 @@
-version=2.42.0
+version=2.43.0
