+++ brevbaker/api-model-common/src/main/kotlin/no/nav/pensjon/brev/api/model/maler/LetterRequest.kt
@@ -39,6 +39,7 @@ class BestillRedigertBrevRequest<T : Brevkode<T>>(
     val letterMarkup: LetterMarkup,
     val alltidValgbareVedlegg: List<AlltidValgbartVedleggKode>,
     val redigerteVedlegg: Map<BrevbakerType.VedleggId, LetterMarkup.Attachment>,
+    val medPDFVedlegg: Boolean
 ) : BrevRequest<T> {
     override fun equals(other: Any?): Boolean {
         if (other !is BestillRedigertBrevRequest<*>) return false
