+++ brevbaker/core/src/main/kotlin/no/nav/brev/brevbaker/BrevbakerPDF.kt
@@ -21,7 +21,7 @@ internal class BrevbakerPDF(
     private val pdfByggerService: PDFByggerService,
     private val pdfVedleggAppender: PDFVedleggAppender,
 ) {
-    suspend fun renderPDF(letter: Letter<BrevbakerBrevdata>, redigertBrev: LetterMarkup? = null, redigerteVedlegg: Map<VedleggId, LetterMarkup.Attachment> = emptyMap()): LetterResponse =
+    suspend fun renderPDF(letter: Letter<BrevbakerBrevdata>, redigertBrev: LetterMarkup? = null, redigerteVedlegg: Map<VedleggId, LetterMarkup.Attachment> = emptyMap(), medPDFVedlegg: Boolean = true): LetterResponse =
         renderCompleteMarkup(letter, redigertBrev, redigerteVedlegg).let { markup ->
             pdfByggerService.producePDF(
                 PDFRequest(
@@ -33,6 +33,7 @@ internal class BrevbakerPDF(
                 ),
             )
         }.let { pdf ->
+            if (!medPDFVedlegg) return@let pdf
             pdfVedleggAppender.leggPaaVedlegg(
                 pdf,
                 letter.template.pdfAttachments
