+++ brevbaker/dsl/src/main/kotlin/no/nav/pensjon/brev/template/vedlegg/PDFTemplate.kt
@@ -6,6 +6,7 @@ import no.nav.pensjon.brev.template.LanguageSupport
 import no.nav.pensjon.brev.template.TextElement
 import no.nav.pensjon.brev.template.dsl.PlainTextOnlyScope
 import no.nav.pensjon.brevbaker.api.model.BrevbakerFelles
+import no.nav.pensjon.brevbaker.api.model.PDFVedlegg
 import no.nav.pensjon.brevbaker.api.model.PDFVedleggData
 import java.util.Objects
