+++ gradle/libs.versions.toml
@@ -5,7 +5,7 @@ ktorVersion = "3.5.1"
 assertJVersion = "3.27.7"
 classGraphVersion = "4.8.184"
 commonsCodecVersion = "1.22.0"
-commonVersion = "2.42.0"
+commonVersion = "2.43.0"
 exposedVersion = "1.3.1"
 flywayVersion = "12.11.0"
 gradleNodePluginVersion = "7.1.0"
